# api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
...
  /api/plant:
    get:
      tags:
      - plant
      operationId: getAllPlants
      parameters:
      - name: deleted
        in: query
        required: false
        schema:
          type: boolean
      - name: page
        in: query
        required: true
        schema:
          minimum: 0
          type: integer
          format: int32
      - name: pageSize
        in: query
        required: true
        schema:
          maximum: 100
          minimum: 5
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagePlant'
    put:
      tags:
      - plant
      operationId: updatePlant
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Plant'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Plant'
    post:
      tags:
      - plant
      operationId: createPlant
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePlant'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Plant'
  /api/photo:
    put:
      tags:
      - photo
      operationId: updatePhoto
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePhoto'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PhotoThumbnail'
    post:
      tags:
      - photo
      operationId: createPhoto
      requestBody:
        content:
          application/json:
            schema:
              required:
              - createPhotoSO
              - photo
              type: object
              properties:
                photo:
                  type: string
                  format: binary
                createPhotoSO:
                  $ref: '#/components/schemas/CreatePhoto'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PhotoThumbnail'
  /api/location:
    get:
      tags:
      - location
      operationId: getAllLocations
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Location'
    put:
      tags:
      - location
      operationId: updateLocation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Location'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Location'
    post:
      tags:
      - location
      operationId: createLocation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLocation'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Location'
  /api/event:
    put:
      tags:
      - event
      operationId: updateEvent
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEvent'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Event'
    post:
      tags:
      - event
      operationId: createEvent
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEvent'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Event'
...
  /api/auth:
    post:
      tags:
      - authenticate
      operationId: authenticateUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Auth'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Token'
...
  /api/plant/{id}:
    get:
      tags:
      - plant
      operationId: getOnePlant
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Plant'
    delete:
      tags:
      - plant
      operationId: delete
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: integer
                format: int64
...
  /api/plant/by-location:
    get:
      tags:
      - plant
      operationId: getAllPlantsByLocation
      parameters:
      - name: locationId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: deleted
        in: query
        required: false
        schema:
          type: boolean
      - name: page
        in: query
        required: true
        schema:
          minimum: 0
          type: integer
          format: int32
      - name: pageSize
        in: query
        required: true
        schema:
          maximum: 100
          minimum: 5
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagePlant'
  /api/photo/{id}:
    get:
      tags:
      - photo
      operationId: getOnePhoto
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Photo'
    delete:
      tags:
      - photo
      operationId: deletePhoto
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: integer
                format: int64
  /api/photo/all/{plantId}:
    get:
      tags:
      - photo
      operationId: getAllPhotoThumbnailsByPlantId
      parameters:
      - name: plantId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: page
        in: query
        required: true
        schema:
          minimum: 0
          type: integer
          format: int32
      - name: pageSize
        in: query
        required: true
        schema:
          maximum: 100
          minimum: 5
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagePhotoThumbnail'
  /api/location/{id}:
    get:
      tags:
      - location
      operationId: getOneLocation
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Location'
    delete:
      tags:
      - location
      operationId: deleteLocation
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: integer
                format: int64
  /api/event/{id}:
    get:
      tags:
      - event
      operationId: getOneEvent
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Event'
    delete:
      tags:
      - event
      operationId: deleteEvent
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
...
  /api/event/all/{plantId}:
    get:
      tags:
      - event
      operationId: getAllEventsByPlantId
      parameters:
      - name: plantId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: page
        in: query
        required: true
        schema:
          minimum: 0
          type: integer
          format: int32
      - name: pageSize
        in: query
        required: true
        schema:
          maximum: 100
          minimum: 5
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageEvent'
components:
  schemas:
...
    EventType:
      required:
      - id
      type: object
      properties:
        id:
          type: integer
          format: int64
        code:
          type: string
        schedulable:
          type: boolean
    PlantBasic:
      required:
      - id
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          maxLength: 100
          minLength: 0
          type: string
...
    LocationBasic:
      required:
      - id
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          maxLength: 80
          minLength: 0
          type: string
    Plant:
      required:
      - deleted
      - id
      - name
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          maxLength: 100
          minLength: 0
          type: string
        description:
          maxLength: 1000
          minLength: 0
          type: string
        type:
          $ref: '#/components/schemas/PlantType'
        location:
          $ref: '#/components/schemas/LocationBasic'
        deleted:
          type: boolean
    PlantType:
      required:
      - id
      type: object
      properties:
        id:
          type: integer
          format: int64
        code:
          maxLength: 80
          minLength: 0
          type: string
    UpdatePhoto:
      required:
      - id
      type: object
      properties:
        id:
          type: integer
          format: int64
        description:
          maxLength: 200
          minLength: 0
          type: string
        plant:
          $ref: '#/components/schemas/PlantBasic'
    PhotoThumbnail:
      required:
      - id
      type: object
      properties:
        id:
          type: integer
          format: int64
        description:
          maxLength: 200
          minLength: 0
          type: string
        thumbnail:
          type: array
          items:
            type: string
            format: byte
        plant:
          $ref: '#/components/schemas/PlantBasic'
    Location:
      required:
      - id
      - name
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          maxLength: 80
          minLength: 0
          type: string
    UpdateEvent:
      required:
      - dateTime
      - id
      - type
      type: object
      properties:
        id:
          type: integer
          format: int64
        type:
          $ref: '#/components/schemas/EventType'
        dateTime:
          type: string
          format: date-time
        description:
          maxLength: 1000
          minLength: 0
          type: string
    Event:
      required:
      - dateTime
      - id
      - plant
      - type
      type: object
      properties:
        id:
          type: integer
          format: int64
        type:
          $ref: '#/components/schemas/EventType'
        dateTime:
          type: string
          format: date-time
        description:
          maxLength: 1000
          minLength: 0
          type: string
        plant:
          $ref: '#/components/schemas/PlantBasic'
...
    CreatePlant:
      required:
      - name
      type: object
      properties:
        name:
          maxLength: 100
          minLength: 0
          type: string
        description:
          maxLength: 1000
          minLength: 0
          type: string
        type:
          $ref: '#/components/schemas/PlantType'
        location:
          $ref: '#/components/schemas/LocationBasic'
    CreatePhoto:
      required:
      - plant
      type: object
      properties:
        description:
          maxLength: 200
          minLength: 0
          type: string
        plant:
          $ref: '#/components/schemas/PlantBasic'
    CreateLocation:
      required:
      - name
      type: object
      properties:
        name:
          maxLength: 80
          minLength: 0
          type: string
    CreateEvent:
      required:
      - dateTime
      - plant
      - type
      type: object
      properties:
        type:
          $ref: '#/components/schemas/EventType'
        dateTime:
          type: string
          format: date-time
        description:
          maxLength: 1000
          minLength: 0
          type: string
        plant:
          $ref: '#/components/schemas/PlantBasic'
    Auth:
      required:
      - password
      - username
      type: object
      properties:
        username:
          maxLength: 25
          minLength: 0
          type: string
        password:
          maxLength: 255
          minLength: 0
          type: string
    Token:
      type: object
      properties:
        type:
          type: string
        token:
          type: string
...
    Pageable:
      type: object
      properties:
        paged:
          type: boolean
        unpaged:
          type: boolean
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        offset:
          type: integer
          format: int64
        sort:
          $ref: '#/components/schemas/Sort'
    Sort:
      type: object
      properties:
        sorted:
          type: boolean
        unsorted:
          type: boolean
        empty:
          type: boolean
    PagePlant:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        number:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        numberOfElements:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/Plant'
        sort:
          $ref: '#/components/schemas/Sort'
        first:
          type: boolean
        last:
          type: boolean
        pageable:
          $ref: '#/components/schemas/Pageable'
        empty:
          type: boolean
    Photo:
      required:
      - id
      type: object
      properties:
        id:
          type: integer
          format: int64
        description:
          maxLength: 200
          minLength: 0
          type: string
        data:
          type: array
          items:
            type: string
            format: byte
        plant:
          $ref: '#/components/schemas/PlantBasic'
    PagePhotoThumbnail:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        number:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        numberOfElements:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/PhotoThumbnail'
        sort:
          $ref: '#/components/schemas/Sort'
        first:
          type: boolean
        last:
          type: boolean
        pageable:
          $ref: '#/components/schemas/Pageable'
        empty:
          type: boolean
    PageEvent:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        number:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        numberOfElements:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/Event'
        sort:
          $ref: '#/components/schemas/Sort'
        first:
          type: boolean
        last:
          type: boolean
        pageable:
          $ref: '#/components/schemas/Pageable'
        empty:
          type: boolean